feat(systemd-resolved): enable multicast DNS resolving
Initially I added the Avahi module for this but that is no longer
necessary. By default systemd-resolved enables mDNS but it's configured
for sending and receiving. That is a potential privacy and security
issue. I do want to be able to use it for things like Chromecast
discovery so it defaults to resolving only.