fix(deploy): commit changes to current branch instead
Committing the changes to another branch and switching back means the
newly added keys are not present in the .sops.yaml file in the branch
that we are working in.

There isn't really a need to push the changes to another branch or the
remote. I'd rather commit changes into the host branch I'm currently
working on so as to keep these changes together in their own PR.